+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This changelog suppose to follow rules defined in the [changelog.md](https://cha
 ## 0.2.0 - TBD
 
 - **Changed**: Protocol changes
+- **Changed**: Flexible database structure
 
 ## 0.1.0 - 2024-06-10